Revert "Merge remote-tracking branch 'upstream/flutter2' into Amir/fix_dpo_indicator_constructor"
This reverts commit 75cdf06f9e797840eeb17b510b7c342d90ba9485.

+++ packages/deriv_technical_analysis/lib/src/indicators/calculations/statistics/variance_indicator.dart
@@ -32,8 +32,7 @@ class VarianceIndicator<T extends IndicatorResult> extends CachedIndicator<T> {
     final double average = _sma.getValue(index).quote;
 
     for (int i = startIndex; i <= index; i++) {
-      final double pow =
-          math.pow(indicator.getValue(i).quote - average, 2) as double;
+      final double pow = math.pow(indicator.getValue(i).quote - average, 2);
       variance = variance + pow;
     }
